Refactored FluidHelper into base FluidField class
FieldHelper had some shared code that both ScalarField and VectorField
used. Instead of keeping it as a static method, I converted it into a
base class with more shared functionality.

I also started using the size of the actual grid instead of the size of
the grid inside of the boundary. As part of that I also started removing
some of the macros, which I think unnecessarily make the code harder to
follow and easier to cause bugs.  I'll clean that up some more later.